% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIQ-SIR-test1.R
\name{test_stat_separate}
\alias{test_stat_separate}
\title{Test statistic computation for hypothesis testing}
\usage{
test_stat_separate(Y, X, taus, m, test_num)
}
\arguments{
\item{Y}{n*1 vector, the observed outcome for inference}

\item{X}{n*p matrix, the observed covariates for inference}

\item{taus}{k*1 vector, a grid of quantile levels; e.g., 0.5 for the median, 0.75 for the 3rd quartile; default is c(0.1, 0.25, 0.5, 0.75, 0.9)}

\item{m}{numeric variable, the order of B-spline function; default is 3}

\item{test_num}{a vector, representing the test corresponds to which covariate(s) in X.}
}
\value{
test statistics for method 'Pearson'
}
\description{
Test statistic computation for hypothesis testing
}
